dealer vehicle status list

+++ database/migrations/2022_11_20_180018_create_dealers_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('dealer_id');
             $table->string('name')->nullable();
-            $table->integer('phone_number')->unsigned()->nullable();
+            $table->bigInteger('phone_number')->unsigned()->nullable();
             $table->timestamps();
         });
     }
